changed ints to floats, adjusted links

+++ PythonFiles/Quadratic_Formula.py
@@ -4,9 +4,9 @@
 import math
 
 print("Input A, B, and C for Ax^2+Bx+C=0")
-A= int(input("A: "))
-B= int(input("B: "))
-C= int(input("C: "))
+A= float(input("A: "))
+B= float(input("B: "))
+C= float(input("C: "))
 D= (-B+math.sqrt(B**2-4*A*C))/(2*A)
 E= (-B-math.sqrt(B**2-4*A*C))/(2*A)
 print("The solutions to the equation " + str(A) + "x^2+" + str(B) + "x+" + str(C) + " are: " + str(D) + " and " + str(E))
